+++ src/kernel/arch/x86/interupt/irq.h
@@ -2,9 +2,10 @@
 #define IRQ_H
 
 #include <stdint.h>
+#include "isr.h"
 
 void irq_init(void);
-void irq_handler(uint32_t irq);
+void irq_handler(uint8_t irq);
 void irq_install(uint8_t irq, void (*handler)(void));
 
 #endif
